backend/requirements.txt
# CPU-only versions to reduce size
torch==2.1.0+cpu
torchvision==0.16.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html

faiss-cpu==1.8.0.post1
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.40.1
huggingface_hub==0.33.4
langchain_community==0.2.19
sentence_transformers==3.2.1